refactored into cogs based on commands/purpose
